Refactor: fitness API route and remove unused object
